refactor: split memory state from matrix ISA
