+++ aten/plena/__init__.py
@@ -3,6 +3,7 @@
 from compiler.aten.plena.compiler import PlenaCompiler
 from compiler.aten.plena.constants import BLEN, IMM2_BOUND, MLEN
 from compiler.aten.plena.isa_compiler import IsaCompiler
+from compiler.aten.plena.memory_state import MemoryStateMixin
 from compiler.aten.plena.tile_compiler import TileCompiler
 from compiler.aten.plena.vars import FPVar, InputVar, TensorVar, VRAMMatrixVar
 
@@ -13,6 +14,7 @@ __all__ = [
     "FPVar",
     "InputVar",
     "IsaCompiler",
+    "MemoryStateMixin",
     "PlenaCompiler",
     "TensorVar",
     "TileCompiler",
