+++ aten/plena_compiler.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 from compiler.aten.plena.compiler import PlenaCompiler
 from compiler.aten.plena.constants import BLEN, IMM2_BOUND, MLEN
 from compiler.aten.plena.isa_compiler import IsaCompiler
+from compiler.aten.plena.memory_state import MemoryStateMixin
 from compiler.aten.plena.tile_compiler import TileCompiler
 from compiler.aten.plena.vars import FPVar, InputVar, TensorVar, VRAMMatrixVar
 
@@ -15,6 +16,7 @@ __all__ = [
     "FPVar",
     "InputVar",
     "IsaCompiler",
+    "MemoryStateMixin",
     "PlenaCompiler",
     "TensorVar",
     "TileCompiler",
